Fix up tests.
